+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All Notable changes to `yaml-front-matter` will be documented in this file
 
+## 2.0.0 - 2017-08-21
+- Renamed `Parser` class to `FrontMatter`
+- `parse` and `parseFile` are now static methods
+- `Document` now has a `__get` implementation for retrieving front matter
+
 ## 1.1.0 - 2017-07-12
 - Added `parseFile`